+++ crates/ruma-client-api/CHANGELOG.md
@@ -10,7 +10,7 @@ Improvements:
 
 * Add support for refresh tokens (MSC2918)
 * Add `ErrorKind::{UnableToAuthorizeJoin, UnableToGrantJoin}` encountered for restricted rooms
-* Add unstable support for timestamp massaging (MSC3316)
+* Add support for timestamp massaging (MSC3316)
 * Add support for querying relating events (MSC2675)
 * Move `filter::RelationType` to `ruma_common::events::relations`
 * Add unstable support for discovering an OpenID Connect server (MSC2965)